refactor: complete constant consolidation and fix build
This commit finalizes the constant consolidation as requested in the pull request review.

- All remaining constants from the legacy `utils/constants.ts` file have been moved to the centralized `lib/constants/index.ts`.
- The now-empty `utils/constants.ts` file has been deleted.
- All import paths throughout the codebase have been updated to point to the new, single source of truth for constants.
- Fixed the resulting build failures by correcting all broken import statements.

This change completes the "Constants & Configuration" phase of the AI Slop Reduction roadmap, resulting in a cleaner, more maintainable, and more consistent codebase. All CI checks are now passing.
